Reorganize the way we document our integrations
We now have one menu item for each kind of integration. The most
important features of each of them have their own submenus as well.

The "integrations" page now shows a link for each integration and the
common features for all of them.

It also improves the introduction to mention test tokens for OAuth2
and add more clues about the process. The idea is to show how to use
the test token and also shows why you need the OAuth2 three steps.