Improve the documentation to clarify REST hooks
We were presenting two different concepts to our developers: webhooks
and RESThooks. This commit changes that to present only webhooks but
reference to RESThooks if the developer wants to read more about it.

Everything is now implemented by Doist but the OAuth implementation of
webhooks is based on RESThooks.

This commit also adds a note to each outgoing webhook to avoid confusion
when dealing with the channel, thread, and OAuth webhooks.